fix: GTP engine path resolution and spdlog enum warning
- Fix findExecutable to return paths relative to workDir since Process
  will chdir to the engine's path before executing
- Cast KeyIdentifier enum to int when logging to avoid spdlog deprecation warning

+++ src/ElementGame.cpp
@@ -176,7 +176,7 @@ void ElementGame::ProcessEvent(Rml::Event& event)
 
     if (event == "keydown" || event == "keyup") {
         Rml::Input::KeyIdentifier key_identifier = (Rml::Input::KeyIdentifier) event.GetParameter< int >("key_identifier", 0);
-        spdlog::debug("ElementGame received {} key={}", event.GetType().c_str(), key_identifier);
+        spdlog::debug("ElementGame received {} key={}", event.GetType().c_str(), static_cast<int>(key_identifier));
         control.keyPress(key_identifier, 0, 0, event == "keydown");
     }
     else if (event == "mousemove") {
